Improved speed of build parser
