+++ src/main/scala/core/bigrammar/GrammarPath.scala
@@ -1,12 +1,10 @@
 package core.bigrammar
 
 import core.bigrammar.grammars._
-import core.bigrammar.printer.UndefinedDestructuringValue
 import core.deltas.node.{Key, NodeField}
 import util.{ExtendedType, GraphBasics, Property}
 
 import scala.collection.mutable
-import scala.util.hashing.Hashing
 
 object GrammarPath {
   val cache: mutable.Map[Class[_], List[Property[BiGrammar, AnyRef]]] = mutable.Map.empty
@@ -59,72 +57,4 @@ trait GrammarPath {
   def descendants: Seq[GrammarReference] = selfAndDescendants.drop(1).collect { case x:GrammarReference => x }
   def selfAndDescendants: Seq[GrammarPath] = GraphBasics.traverseBreadth[GrammarPath](Seq(this),
     path => path.newChildren)
-}
-
-class RootGrammar(val value: BiGrammar) extends GrammarPath
-{
-  override def ancestorGrammars = Set.empty
-
-  override def ancestors: Seq[GrammarPath] = Seq.empty
-
-  override def hashCode(): Int = value.hashCode()
-
-  override def equals(obj: Any): Boolean = obj match {
-    case other: RootGrammar => value.equals(other.value)
-    case _ => false
-  }
-}
-
-class GrammarReference(val previous: GrammarPath, val property: Property[BiGrammar, AnyRef]) extends GrammarPath
-{
-  def parent: BiGrammar = previous.value
-  private var cachedValue: BiGrammar = _
-  private var cachedHashCode: Option[Int] = None
-  private var cachedAncestorGrammars: Set[BiGrammar] = _
-
-  def ancestorGrammars: Set[BiGrammar] = {
-    if (cachedAncestorGrammars == null)
-      cachedAncestorGrammars = previous.ancestorGrammars + parent
-    cachedAncestorGrammars
-  }
-
-  def value: BiGrammar = {
-    if (cachedValue == null)
-      cachedValue = property.get(parent).asInstanceOf[BiGrammar]
-    cachedValue
-  }
-
-  def set(newValue: BiGrammar): Unit = {
-    property.set(parent, newValue)
-    cachedValue = null
-    cachedAncestorGrammars = null
-    cachedHashCode = None
-  }
-
-  override def hashCode(): Int = {
-    if (cachedHashCode.isEmpty)
-      cachedHashCode = Some(Hashing.default.hash((previous.hashCode(), property.hashCode())))
-    cachedHashCode.get
-  }
-
-  override def equals(obj: scala.Any): Boolean = obj match {
-    case other: GrammarReference =>
-      other.property.equals(property) && other.previous.equals(previous)
-    case _ => false
-  }
-
-  def removeMeFromOption(): Unit = {
-    val choiceParent = parent.asInstanceOf[Choice]
-    val me = value
-    val sibling = Set(choiceParent.left,choiceParent.right).filter(grammar => grammar != me).head
-    previous.asInstanceOf[GrammarReference].set(sibling)
-  }
-
-  def removeMe(): Unit = {
-    set(ValueGrammar(UndefinedDestructuringValue)) //TODO add transformation to remove ValueGrammar(Unit)
-  }
-
-  override def toString = s"$value <INSIDE> $parent"
-
-  override def ancestors: Seq[GrammarPath] = Seq(previous) ++ previous.ancestors
 }
